Move component out of app and fix some names
